[chore] move action types to the types folder
