+++ src/types/index.js
@@ -1,6 +1,10 @@
+// page actions
+export const TITLE_CHANGE = 'TITLE_CHANGE';
+
+// user info actions
 export const FETCH_USER_INFO = 'FETCH_USER_INFO';
 
-// my activities
+// my activities actions
 export const MY_ACTIVITIES_GET_REQUEST = 'MY_ACTIVITIES_GET_REQUEST';
 export const MY_ACTIVITIES_GET_SUCCESS = 'MY_ACTIVITIES_GET_SUCCESS';
 export const MY_ACTIVITIES_GET_FAILURE = 'MY_ACTIVITIES_GET_FAILURE';
