--- src/types/pageActionTypes.js
@@ -1,2 +1,0 @@
-export const TITLE_CHANGE = 'TITLE_CHANGE';
-export default { TITLE_CHANGE };
